+++ src/client/staticdata.h
@@ -356,7 +356,7 @@ struct StoreData
 struct CyclopediaCharacterGeneralStats
 {
     uint64_t experience;
-    uint16_t level;
+    uint32_t level;
     uint8_t levelPercent;
     uint16_t baseExpGain;
     uint16_t lowLevelExpBonus;
